Separate power level sync to separate module
Otherwise things don't work. Add instructions in
readme how to enable it. Add some better logging.
